Pushed v7 layout

+++ modules/Reports/models/ScheduleReports.php
@@ -9,9 +9,8 @@
  * All Rights Reserved.
  * *********************************************************************************** */
 
+require_once 'vtlib/Vtiger/Cron.php';
 class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
-
-	var $scheduledFormat = 'CSV';
 
 	static $SCHEDULED_DAILY = 1;
 	static $SCHEDULED_WEEKLY = 2;
@@ -53,6 +52,7 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 	 */
 	public function saveScheduleReport() {
 		$adb = PearDatabase::getInstance();
+        $scheduleCronTx = true;
 
 		$reportid = $this->get('reportid');
         $scheduleid = $this->get('scheduleid');
@@ -66,28 +66,32 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 		if ($scheduleid == self::$SCHEDULED_ON_SPECIFIC_DATE) {
 			$date = $this->get('schdate');
 			$dateDBFormat = DateTimeField::convertToDBFormat($date);
-            $nextTriggerTime = $dateDBFormat.' '.$schtime;
-            $currentTime = Vtiger_Util_Helper::getActiveAdminCurrentDateTime();
-            if($nextTriggerTime > $currentTime) {
-                $this->set('next_trigger_time', $nextTriggerTime);
-            } else {
-                $this->set('next_trigger_time', date('Y-m-d H:i:s', strtotime('+10 year')));
-            }
+                        $nextTriggerTime = $dateDBFormat.' '.$schtime;
+                        $currentTime = date('Y-m-d H:i:s');
+                        $user = Users::getActiveAdminUser();
+                        $dateTime = new DateTimeField($nextTriggerTime);
+                        $nextTriggerTime = $dateTime->getDBInsertDateTimeValue($user);
+                        if($nextTriggerTime > $currentTime) {
+                            $this->set('next_trigger_time', $nextTriggerTime);
+                        } else {
+                            $this->set('next_trigger_time', date('Y-m-d H:i:s', strtotime('+10 year')));
+                        }
 			$schdate = Zend_Json::encode(array($dateDBFormat));
 		} else if ($scheduleid == self::$SCHEDULED_WEEKLY) {
 			$schdayoftheweek = Zend_Json::encode($this->get('schdayoftheweek'));
-            $this->set('schdayoftheweek', $schdayoftheweek);
+                        $this->set('schdayoftheweek', $schdayoftheweek);
 		} else if ($scheduleid == self::$SCHEDULED_MONTHLY_BY_DATE) {
 			$schdayofthemonth = Zend_Json::encode($this->get('schdayofthemonth'));
-            $this->set('schdayofthemonth', $schdayofthemonth);
+                        $this->set('schdayofthemonth', $schdayofthemonth);
 		} else if ($scheduleid == self::$SCHEDULED_ANNUALLY) {
 			$schannualdates = Zend_Json::encode($this->get('schannualdates'));
-            $this->set('schannualdates', $schannualdates);
+                        $this->set('schannualdates', $schannualdates);
 		}
 
 		$recipients = Zend_Json::encode($this->get('recipients'));
 		$specificemails = Zend_Json::encode($this->get('specificemails'));
 		$isReportScheduled = $this->get('isReportScheduled');
+        $fileFormat = $this->get('fileformat');
 
         if($scheduleid != self::$SCHEDULED_ON_SPECIFIC_DATE) {
             $nextTriggerTime = $this->getNextTriggerTime();
@@ -96,16 +100,23 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 			$deleteScheduledReportSql = "DELETE FROM vtiger_schedulereports WHERE reportid=?";
 			$adb->pquery($deleteScheduledReportSql, array($reportid));
 		} else {
-			$checkScheduledResult = $adb->pquery('SELECT 1 FROM vtiger_schedulereports WHERE reportid=?', array($reportid));
+			$checkScheduledResult = $adb->pquery('SELECT next_trigger_time FROM vtiger_schedulereports WHERE reportid=?', array($reportid));
 			if ($adb->num_rows($checkScheduledResult) > 0) {
-				$scheduledReportSql = 'UPDATE vtiger_schedulereports SET scheduleid=?, recipients=?, schdate=?, schtime=?, schdayoftheweek=?, schdayofthemonth=?, schannualdates=?, specificemails=?, next_trigger_time=? WHERE reportid=?';
-				$adb->pquery($scheduledReportSql, array($scheduleid, $recipients, $schdate, $schtime, $schdayoftheweek, $schdayofthemonth, $schannualdates, $specificemails, $nextTriggerTime, $reportid));
+                $scheduledReportSql = 'UPDATE vtiger_schedulereports SET scheduleid=?, recipients=?, schdate=?, schtime=?, schdayoftheweek=?, schdayofthemonth=?, schannualdates=?, specificemails=?, next_trigger_time=?, fileformat = ? WHERE reportid=?';
+				$adb->pquery($scheduledReportSql, array($scheduleid, $recipients, $schdate, $schtime, $schdayoftheweek, $schdayofthemonth, $schannualdates, $specificemails, $nextTriggerTime, $fileFormat, $reportid));
+                $trigger_time = $adb->query_result($checkScheduledResult,0,'next_trigger_time');
+				if($trigger_time == $nextTriggerTime){
+                    $scheduleCronTx = false;
+                }
 			} else {
-				$scheduleReportSql = 'INSERT INTO vtiger_schedulereports (reportid,scheduleid,recipients,schdate,schtime,schdayoftheweek,schdayofthemonth,schannualdates,next_trigger_time,specificemails) VALUES (?,?,?,?,?,?,?,?,?,?)';
-				$adb->pquery($scheduleReportSql, array($reportid, $scheduleid, $recipients, $schdate, $schtime, $schdayoftheweek, $schdayofthemonth, $schannualdates, $nextTriggerTime, $specificemails));
+				$scheduleReportSql = 'INSERT INTO vtiger_schedulereports (reportid,scheduleid,recipients,schdate,schtime,schdayoftheweek,schdayofthemonth,schannualdates,next_trigger_time,specificemails, fileformat) VALUES (?,?,?,?,?,?,?,?,?,?,?)';
+				$adb->pquery($scheduleReportSql, array($reportid, $scheduleid, $recipients, $schdate, $schtime, $schdayoftheweek, $schdayofthemonth, $schannualdates, $nextTriggerTime,$specificemails,$fileFormat));
 			}
+            }
+            if($scheduleCronTx){
+                $this->scheduleCronTx($nextTriggerTime);
+            }
 		}
-	}
 
 	public function getRecipientEmails() {
 		$recipientsInfo = $this->get('recipients');
@@ -146,6 +157,9 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 				foreach ($recipients['Groups'] as $groupId) {
 					$userGroups = new GetGroupUsers();
 					$userGroups->getAllUsersInGroup($groupId);
+
+					//Clearing static cache for sub groups
+					GetGroupUsers::$groupIdsList = array();
 					$recipientsList = array_merge($recipientsList, $userGroups->group_users);
 				}
 			}
@@ -174,13 +188,14 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 
 	public function sendEmail() {
 		require_once 'vtlib/Vtiger/Mailer.php';
+        $currentUserModel = Users_Record_Model::getCurrentUserModel();
 
 		$vtigerMailer = new Vtiger_Mailer();
 
 		$recipientEmails = $this->getRecipientEmails();
         Vtiger_Utils::ModuleLog('ScheduleReprots', $recipientEmails);
 		foreach ($recipientEmails as $name => $email) {
-			$vtigerMailer->AddAddress($email, $name);
+			$vtigerMailer->AddAddress($email, decode_html($name));
 		}
 		vimport('~modules/Report/models/Record.php');
 		$reportRecordModel = Reports_Record_Model::getInstanceById($this->get('reportid'));
@@ -189,34 +204,59 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 		$reportname = decode_html($reportRecordModel->getName());
         $subject = $reportname;
         Vtiger_Utils::ModuleLog('ScheduleReprot Name ::', $reportname);
+        if(empty($vtigerMailer->From)) {
+            $fromName = trim($currentUserModel->get('first_name').' '.$currentUserModel->get('last_name'));
+            $vtigerMailer->ConfigSenderInfo($currentUserModel->get('email1'), $fromName);
+        }
 		$vtigerMailer->Subject = $subject;
 		$vtigerMailer->Body = $this->getEmailContent($reportRecordModel);
+        $plainBody = decode_html($vtigerMailer->Body);
+        $plainBody = preg_replace(array("/<p>/i","/<br>/i","/<br \/>/i"),array("\n","\n","\n"),$plainBody);
+        $plainBody = strip_tags($plainBody);
+        $plainBody = Emails_Mailer_Model::convertToAscii($plainBody);
+        $vtigerMailer->AltBody = $plainBody;
+
 		$vtigerMailer->IsHTML();
 
-		$baseFileName = $reportname . '_' . $currentTime;
-
+        $baseFileName = preg_replace("/[^\p{L}\p{N}\s]+/", "", $reportname);
+        
 		$oReportRun = ReportRun::getInstance($this->get('reportid'));
-		$reportFormat = $this->scheduledFormat;
+		$reportFormat = $this->get('fileformat');
 		$attachments = array();
+        $reportType = $reportRecordModel->get('reporttype');
 
-		if ($reportFormat == 'CSV') {
-			$fileName = $baseFileName . '.csv';
-			$filePath = 'storage/' . $fileName;
-			$attachments[$fileName] = $filePath;
-			$oReportRun->writeReportToCSVFile($filePath);
-		}
+        if($reportType != 'chart') {
+            if ($reportFormat == 'CSV') {
+                $fileName = $baseFileName . '.csv';
+                $filePath = 'storage/' . $fileName;
+                $attachments[$fileName] = $filePath;
+                if($reportType == 'pivot') {
+                    $oReportRun->writePivotReportToCSVFile($filePath);
+                } else {
+                    $oReportRun->writeReportToCSVFile($filePath);
+                }
+            } else if($reportFormat == 'XLS') {
+                $fileName = $baseFileName . '.xls';
+                $filePath = 'storage/' . $fileName;
+                $attachments[$fileName] = $filePath;
+                if($reportType == 'pivot') {
+                    $oReportRun->writePivotReportToExcelFile($filePath);
+                } else {
+                    $oReportRun->writeReportToExcelFile($filePath);
+                }
+            }
 
-		foreach ($attachments as $attachmentName => $path) {
-			$vtigerMailer->AddAttachment($path, decode_html($attachmentName));
-		}
-		//Added cc to account owner
-		$accountOwnerId = Users::getActiveAdminId();
-		$vtigerMailer->AddCC(getUserEmail($accountOwnerId), getUserFullName($accountOwnerId));
+            foreach ($attachments as $attachmentName => $path) {
+                $vtigerMailer->AddAttachment($path, decode_html($attachmentName));
+            }
+        }
 		$status = $vtigerMailer->Send(true);
 
-		foreach ($attachments as $attachmentName => $path) {
-			unlink($path);
-		}
+        if($reportType != 'chart') {
+            foreach ($attachments as $attachmentName => $path) {
+                unlink($path);
+            }
+        }
 		return $status;
 	}
 
@@ -227,7 +267,7 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 	 */
 	function getNextTriggerTime() {
 		require_once 'modules/com_vtiger_workflow/VTWorkflowManager.inc';
-		$default_timezone = vglobal('default_timezine');
+		$default_timezone = vglobal('default_timezone');
 		$admin = Users::getActiveAdminUser();
 		$adminTimeZone = $admin->time_zone;
 		@date_default_timezone_set($adminTimeZone);
@@ -243,10 +283,6 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 			$nextTime = $workflow->getNextTriggerTimeForWeekly($this->get('schdayoftheweek'), $this->get('schtime'));
 		}
 
-		if ($scheduleType == self::$SCHEDULED_ON_SPECIFIC_DATE) {
-			$nextTime = date('Y-m-d H:i:s', strtotime('+10 year'));
-		}
-
 		if ($scheduleType == self::$SCHEDULED_MONTHLY_BY_DATE) {
 			$nextTime = $workflow->getNextTriggerTimeForMonthlyByDate($this->get('schdayofthemonth'), $this->get('schtime'));
 		}
@@ -255,6 +291,10 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 			$nextTime = $workflow->getNextTriggerTimeForAnnualDates($this->get('schannualdates'), $this->get('schtime'));
 		}
 		@date_default_timezone_set($default_timezone);
+                if($scheduleType != self::$SCHEDULED_ON_SPECIFIC_DATE) {
+                    $dateTime = new DateTimeField($nextTime);
+                    $nextTime = $dateTime->getDBInsertDateTimeValue($admin);
+                }
 		return $nextTime;
 	}
 
@@ -263,27 +303,24 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 		$nextTriggerTime = $this->getNextTriggerTime();
         Vtiger_Utils::ModuleLog('ScheduleReprot Next Trigger Time >> ', $nextTriggerTime);
 		$adb->pquery('UPDATE vtiger_schedulereports SET next_trigger_time=? WHERE reportid=?', array($nextTriggerTime, $this->get('reportid')));
+        $this->scheduleCronTx($nextTriggerTime);
         Vtiger_Utils::ModuleLog('ScheduleReprot', 'Next Trigger Time updated');
 	}
 
 	public static function getScheduledReports() {
 		$adb = PearDatabase::getInstance();
-        $default_timezone = vglobal('default_timezone');
 
-        // set the time zone to the admin's time zone, this is needed so that the scheduled reprots will be triggered
-		// at admin's time zone rather than the systems time zone. This is specially needed for Hourly and Daily scheduled reports
-		$admin = Users::getActiveAdminUser();
-		$adminTimeZone = $admin->time_zone;
-		@date_default_timezone_set($adminTimeZone);
 		$currentTimestamp  = date("Y-m-d H:i:s");
-		@date_default_timezone_set($default_timezone);
-		$result = $adb->pquery("SELECT reportid FROM vtiger_schedulereports WHERE next_trigger_time = '' || next_trigger_time <= ?", array($currentTimestamp));
+		$result = $adb->pquery("SELECT reportid FROM vtiger_schedulereports
+								INNER JOIN vtiger_reportmodules ON vtiger_reportmodules.reportmodulesid = vtiger_schedulereports.reportid
+								INNER JOIN vtiger_tab ON vtiger_tab.name = vtiger_reportmodules.primarymodule AND presence = 0
+								WHERE next_trigger_time <= ? AND next_trigger_time IS NOT NULL", array($currentTimestamp));
 
 		$scheduledReports = array();
 		$noOfScheduledReports = $adb->num_rows($result);
 		for ($i = 0; $i < $noOfScheduledReports; ++$i) {
 			$recordId = $adb->query_result($result, $i, 'reportid');
-			$scheduledReports[] = self::getInstanceById($recordId);
+			$scheduledReports[$recordId] = self::getInstanceById($recordId);
 		}
 		return $scheduledReports;
 	}
@@ -298,11 +335,20 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 		if(empty($current_language)) $current_language = 'en_us';
 
 		$scheduledReports = self::getScheduledReports();
-		foreach ($scheduledReports as $scheduledReport) {
-			$status = $scheduledReport->sendEmail();
+		foreach ($scheduledReports as $reportId => $scheduledReport) {
+            $reportRecordModel = Reports_Record_Model::getInstanceById($reportId);
+            $reportType = $reportRecordModel->get('reporttype');
+            if($reportType == 'pivot' || $reportType == 'chart') {
+                $status = $scheduledReport->sendEmail();
+            } else {
+                $query = $reportRecordModel->getReportSQL();
+                $countQuery = $reportRecordModel->generateCountQuery($query);
+                if($reportRecordModel->getReportsCount($countQuery) > 0){
+                    $status = $scheduledReport->sendEmail();
+                }
+            }
             Vtiger_Utils::ModuleLog('ScheduleReprot Send Mail Status ', $status);
-			if($status)
-				$scheduledReport->updateNextTriggerTime();
+			$scheduledReport->updateNextTriggerTime();
 		}
 		$util->revertUser();
 		return $status;
@@ -380,5 +426,48 @@ class Reports_ScheduleReports_Model extends Vtiger_Base_Model {
 
 	return $body;
 	}
+    
+    public function getNextTriggerTimeInUserFormat() {
+        $dateTime = new DateTimeField($this->get('next_trigger_time'));
+        $nextTriggerTime = $dateTime->getDisplayDateTimeValue();
+        $valueParts = explode(' ', $nextTriggerTime);
+        $value = $valueParts[0].' '.Vtiger_Time_UIType::getDisplayValue($valueParts[1]);
+        return $value;
+    }
+    
+    /**
+     * To Schedule Cron Job in CronTx
+     * @param type $nextTriggerTime
+     */
+    protected function scheduleCronTx($nextTriggerTime){
+        if(!is_array($nextTriggerTime)){
+            $nextTriggerTime = array($nextTriggerTime);
+        }
+        if(Vtiger_Cron::$enableCronTx){
+            $cron = Vtiger_Cron::getInstance('ScheduleReports');
+            $admin = Users::getActiveAdminUser();
+            foreach ($nextTriggerTime as $triggertime){
+                // Converting next trigger time to UTC format for registering to CronTx
+                $dateTime = new DateTime($triggertime, new DateTimeZone($admin->time_zone));
+                $cron->schedule($dateTime->format('Y-m-d H:i:s'));
+            }
+        }
+    }
+    
+    /**
+     * To schedule CronTx for already scheduled reports
+     */
+    public function sendScheduledReportsToCronTx(){
+        $adb = PearDatabase::getInstance();
+        $result = $adb->pquery("SELECT next_trigger_time FROM vtiger_schedulereports",array());
+        $count = $adb->num_rows($result);
+        if($count > 0){
+            $trigger_times = array();
+            while($row = $adb->fetchByAssoc($result)){
+                $trigger_times[] = $row['next_trigger_time'];
+            }
+            $this->scheduleCronTx($trigger_times);
+        }
+    }
 }
